:truck: Update folder name / Delete package
	폴더 이름을 변경하고, 패키지 제이슨을 삭제하였습니다.
