--- client/.eslintrc.js
@@ -1,27 +1,0 @@
-module.exports = {
-    env: {
-        browser: true,
-        es6: true,
-        node: true,
-    },
-    extends: ['eslint:recommended', 'plugin:react/recommended'],
-    globals: {
-        Atomics: 'readonly',
-        SharedArrayBuffer: 'readonly',
-    },
-    parserOptions: {
-        ecmaFeatures: {
-            jsx: true,
-        },
-        ecmaVersion: 2018,
-        sourceType: 'module',
-    },
-    plugins: ['react'],
-    rules: {
-        'no-console': 'off',
-        'no-unused-vars': [
-            'error',
-            { vars: 'all', args: 'after-used', ignoreRestSiblings: false },
-        ],
-    },
-};
