+++ client/README.md
@@ -1,22 +1,1 @@
-# 🎨 색갈피 Client
-
-`React`로 구현된 색갈피의 client 파일입니다. <br />
-색갈피는 한글로 구성된 컬러 팔레트 웹사이트로, 컬러 팔레트의 제작 및 유저들 간의 팔레트 공유가 가능합니다.
-
-## ✨ Features
-
--   [x] 팔레트 조회
--   [x] 팔레트 편집 및 생성
--   [x] 팔레트 공유 및 다운로드
--   [x] 회원가입
--   [x] 로그인
--   [x] 로그아웃
--   [x] 회원 정보 수정
-
-## 📦 Dependencies
-
--   React
--   React-Router
--   React-Hooks
--   SASS
--   Styled-Component
+색갈피!
