+++ client/package.json
@@ -1,18 +1,27 @@
 {
-  "name": "client",
+  "name": "type-client",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.29",
-    "@fortawesome/free-brands-svg-icons": "^5.13.1",
-    "@fortawesome/free-regular-svg-icons": "^5.13.1",
-    "@fortawesome/free-solid-svg-icons": "^5.13.1",
+    "@fortawesome/fontawesome-svg-core": "^1.2.30",
+    "@fortawesome/free-brands-svg-icons": "^5.14.0",
+    "@fortawesome/free-regular-svg-icons": "^5.14.0",
+    "@fortawesome/free-solid-svg-icons": "^5.14.0",
     "@fortawesome/react-fontawesome": "^0.1.11",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
-    "antd": "^4.4.0",
-    "axios": "^0.19.2",
+    "@types/antd": "^1.0.0",
+    "@types/jest": "^24.9.1",
+    "@types/node": "^12.12.55",
+    "@types/node-sass": "^4.11.1",
+    "@types/react": "^16.9.49",
+    "@types/react-color": "^3.0.4",
+    "@types/react-dom": "^16.9.8",
+    "@types/react-redux": "^7.1.9",
+    "@types/styled-components": "^5.1.3",
+    "antd": "^4.6.3",
+    "axios": "^0.20.0",
     "dom-to-image": "^2.6.0",
     "hex-rgb-color-code-converter": "^1.0.2",
     "node-sass": "^4.14.1",
@@ -20,9 +29,12 @@
     "react-color": "^2.18.1",
     "react-copy-to-clipboard": "^5.0.2",
     "react-dom": "^16.13.1",
+    "react-redux": "^7.2.1",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.1",
-    "styled-components": "^5.1.1"
+    "react-scripts": "3.4.3",
+    "redux": "^4.0.5",
+    "styled-components": "^5.2.0",
+    "typescript": "^3.7.5"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -44,5 +56,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@types/react-router-dom": "^5.1.5"
   }
 }
