--- client/src/hook/useWindowSize.js
@@ -1,18 +1,0 @@
-import { useLayoutEffect, useState } from 'react';
-
-// window 창의 리사이징을 검사하는 Hook
-// 출처: https://stackoverflow.com/questions/19014250/rerender-view-on-browser-resize-with-react
-const useWindowSize = () => {
-    const [size, setSize] = useState([0, 0]);
-    useLayoutEffect(() => {
-        const updateSize = () => {
-            setSize([window.innerWidth, window.innerHeight]);
-        };
-        window.addEventListener('resize', updateSize);
-        updateSize();
-        return () => window.removeEventListener('resize', updateSize);
-    }, []);
-    return size;
-};
-
-export default useWindowSize;
