+++ client/src/index.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from './App';
+
+import { createStore } from 'redux';
+import { Provider } from 'react-redux';
+import rootReducer from './reducer';
+
+const store = createStore(rootReducer);
+
+ReactDOM.render(
+  <Provider store={store}>
+    <App />
+  </Provider>,
+  document.getElementById('root')
+);
+
+ReactDOM.render(<App />, document.getElementById('root'));
