+++ client/src/pages/Main.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Main() {
+  return <div>ㄴㄴ</div>;
+}
+
+export default Main;
