+++ client/src/reducer/index.ts
@@ -1,0 +1,10 @@
+import { combineReducers } from 'redux';
+import paletteReducer from './paletteReducer';
+
+const rootReducer = combineReducers({
+  paletteReducer,
+});
+
+export default rootReducer;
+
+export type RootState = ReturnType<typeof rootReducer>;
