+++ client/src/scss/config/App.scss
@@ -16,46 +16,46 @@
 @import '../templetes/Main.scss';
 
 html {
-    body {
-        background-color: $main;
+  body {
+    background-color: $main;
+  }
+  main {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 80%;
+    margin: 0 auto;
+    min-height: 700px;
+    h1 {
+      font-size: 3em;
     }
-    main {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        width: 80%;
-        margin: 0 auto;
-        min-height: 700px;
-        h1 {
-            font-size: 3em;
-        }
-    }
+  }
 
-    //유저 페이지의 기본 wrapper입니다
-    .userPageWrapper {
-        width: 900px;
-        text-align: center;
-        display: flex;
-        justify-content: center;
-    }
+  //유저 페이지의 기본 wrapper입니다
+  .userPageWrapper {
+    width: 900px;
+    text-align: center;
+    display: flex;
+    justify-content: center;
+  }
 
-    .h1 {
-        font-size: 3em;
-        font-weight: 900;
-        margin-left: 30px;
-    }
+  .h1 {
+    font-size: 3em;
+    font-weight: 900;
+    margin-left: 30px;
+  }
 
-    //드래그 색상 바꿨습니다.
-    ::selection {
-        background-color: #1f5bb5;
-        color: #f6f8fa;
-    }
+  //드래그 색상 바꿨습니다.
+  ::selection {
+    background-color: #1f5bb5;
+    color: #f6f8fa;
+  }
 
-    .no-drag {
-        -ms-user-select: none;
-        -moz-user-select: -moz-none;
-        -webkit-user-select: none;
-        -khtml-user-select: none;
-        user-select: none;
-    }
+  .no-drag {
+    -ms-user-select: none;
+    -moz-user-select: -moz-none;
+    -webkit-user-select: none;
+    -khtml-user-select: none;
+    user-select: none;
+  }
 }
